@@ -7,9 +7,12 @@ package dachman.lucas.letsgoapp2;
  * not used for final product
  */
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+
 import android.os.Bundle;
 import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
+
 import dachman.lucas.letsgoapp2.R;
 import dachman.lucas.letsgoapp2.LoginActivity;
 
