@@ -1,8 +1,10 @@
 package dachman.lucas.letsgoapp2;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+
 import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 /**
  * @author jasmine
